changed int data type to double

+++ final/Foundation4/Biking.cs
@@ -1,21 +1,22 @@
 public class Biking : Activity
 {
-    private int _speed ;
+    private double _speed ;
     public Biking (string date, int length, int counter, int speed) : base (date, length, counter)
     {
         _speed = speed;
     }
-    public override int DistanceCalculation()
+    public override double DistanceCalculation()
     {
         // (speed * minutes) * 60
         return ((_speed * _length) / 60);
     }
-    public override int SpeedCalculation()
+    public override double SpeedCalculation()
     {
         return _speed;
     }
-    public override int PaceCalculation()
+    public override double PaceCalculation()
     {
+        // doesn't appear
         // Pace (min per mile or min per km)= minutes / distance
         return (_length / ((_speed * _length)));
     }
